requirements.txt
keras==3.4.1
numpy==1.26.4
pillow==10.4.0
tensorflow==2.17.0
tensorflow-hub==0.16.1